requirements.txt
requests
pylint
flake8
PyGithub
bandit
python-dotenv
openai==0.28
anthropic
